:bug: fix bug because on repository  we got long type, on controller and upper than repository we got int

+++ src/main/java/com/esgi/api_project_annuel/Domain/repository/UserRoleGroupRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 @Repository
 public interface UserRoleGroupRepository extends JpaRepository<UserRoleGroup, Integer> {
-    UserRoleGroup findById(long id);
+    UserRoleGroup findById(int id);
 
     UserRoleGroup findByGroupAndUser(Group group, User user);
